# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users": {
      "get": {
        "operationId": "UsersController_getUsers",
        "summary": "Get all users",
        "description": "Retrieve all users. Requires admin authentication.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Users retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UsersController_getUser",
        "summary": "Get user profile",
        "description": "Retrieve user profile information by user ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID (UUID)",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "123e4567-e89b-12d3-a456-426614174000",
                    "email": "user@example.com",
                    "name": "John Doe",
                    "avatar": "https://github.com/username.png",
                    "department": "Computer Science",
                    "grade": 3,
                    "bio": "Full-stack developer",
                    "theme": "dark",
                    "language": "ko",
                    "createdAt": "2024-01-01T00:00:00.000Z",
                    "updatedAt": "2024-01-01T00:00:00.000Z"
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_updateUser",
        "summary": "Update user profile",
        "description": "Update user profile information. Users can only update their own profile.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID (must match authenticated user)",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User profile updated successfully"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token"
          },
          "403": {
            "description": "Forbidden - Cannot update other users profiles"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/users/{id}/projects": {
      "get": {
        "operationId": "UsersController_getUserProjects",
        "summary": "Get user projects",
        "description": "Retrieve all projects where the user is a member",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User projects retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": "456e7890-e12b-34d5-a678-426614174111",
                      "name": "E-Commerce Platform",
                      "slug": "ecommerce-platform",
                      "description": "Full-stack e-commerce solution",
                      "visibility": "PUBLIC",
                      "role": "OWNER",
                      "createdAt": "2024-01-01T00:00:00.000Z"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "WKU Software Crew API",
    "description": "RESTful API for WKU Software Crew platform - A comprehensive learning and collaboration platform for software development students",
    "version": "1.0",
    "contact": {
      "name": "WKU Software Crew",
      "url": "https://github.com/wku-crew",
      "email": "contact@wku-crew.com"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "User display name",
            "example": "John Doe"
          },
          "bio": {
            "type": "string",
            "description": "User biography or introduction",
            "example": "Full-stack developer passionate about open source"
          },
          "avatar": {
            "type": "string",
            "description": "Profile picture URL",
            "example": "https://github.com/username.png"
          },
          "department": {
            "type": "string",
            "description": "University department",
            "example": "Computer Science"
          },
          "grade": {
            "type": "number",
            "description": "Academic grade (1-6)",
            "example": 3,
            "minimum": 1,
            "maximum": 6
          },
          "theme": {
            "type": "string",
            "description": "UI theme preference",
            "example": "dark",
            "enum": [
              "light",
              "dark"
            ]
          },
          "language": {
            "type": "string",
            "description": "Preferred language",
            "example": "ko",
            "enum": [
              "ko",
              "en"
            ]
          }
        }
      },
...
    }
  }
}